Is your feature request related to a problem? Please describe.
When moving data between instances in some cases not all elements exist in the dimensions of the cube on the target instance. I know this and accept this will result in a loss of data. Loss of data is already accepted by checking the cellisupdatable. In the current process, where the data point does not exist for a record, an entry is added to the process log and the process completes with minor errors. I would prefer that the process skips those records that do not have a data point and complete successfully.

Describe the solution you'd like
An additional parameter, 'pStrict' with the default value 1. If pStrict = 1 then the current behaviour is maintained. If pStrict = 0 then the element is for the dimension is checked. If the element does not exist in any dimension the record is skipped.

Describe alternatives you've considered
Due the number records and different target instance the checking needs to be handled in the }bedrock process.